Support fractional scaling
Change scale factor from int to float

+++ libqtile/backend/wayland/drawer.py
@@ -92,7 +92,9 @@ class Drawer(drawer.Drawer):
             context.fill()
 
         damage = PixmanRegion32()
-        damage.init_rect(offsetx * scale, offsety * scale, width * scale, height * scale)
+        damage.init_rect(
+            int(offsetx * scale), int(offsety * scale), int(width * scale), int(height * scale)
+        )
         # TODO: do we really need to `set_buffer` here? would be good to just set damage
         self._win._scene_buffer.set_buffer_with_damage(self._win.wlr_buffer, damage)
         damage.fini()
